backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
openai==1.98.0
langchain==0.0.340
langchain-community==0.0.2
langchain-openai==0.0.5
soundfile==0.12.1
ffmpeg-python==0.2.0
aiofiles==23.2.1
jinja2==3.1.2
yt-dlp==2023.12.30